pyproject.toml
[project]
name = "sandraetl" 
version = "0.1.0"
description = "A Python ELT pipeline for extracting, transforming, and loading data."
authors = [{name = "Sandra", email = "kiptumjsandra@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "pandas",
    "sqlalchemy",
    "psycopg2-binary",
    "requests",
    "numpy"
]


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sandraetl"]
